pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Pareiko Vladislav <parieiko.vlad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
pymysql = "^1.1.1"
mysqlclient = "^2.2.4"
alembic = "^1.13.2"
fastapi-async-sqlalchemy = "^0.6.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
